refactor: fix cyclic type dependency
