+++ packages/analytics-js/__tests__/app/RudderAnalytics.test.ts
@@ -1,5 +1,5 @@
 import type { LoadOptions } from '@rudderstack/analytics-js-common/types/LoadOptions';
-import type { RSACustomIntegration } from '@rudderstack/analytics-js-common/types/CustomIntegration';
+import type { RSACustomIntegration } from '@rudderstack/analytics-js-common/types/IRudderAnalytics';
 import { resetState, state } from '../../src/state';
 import { RudderAnalytics } from '../../src/app/RudderAnalytics';
 import { Analytics } from '../../src/components/core/Analytics';
